+++ src/Regressors/MLPRegressor.php
@@ -43,6 +43,7 @@ use const Rubix\ML\EPSILON;
  *
  * References:
  * [1] G. E. Hinton. (1989). Connectionist learning procedures.
+ * [2] L. Prechelt. (1997). Early Stopping - but when?
  *
  * @category    Machine Learning
  * @package     Rubix/ML
@@ -114,7 +115,7 @@ class MLPRegressor implements Estimator, Online, Verbose, Persistable
     protected $holdout;
 
     /**
-     * The number of epochs without improvement in the training loss to wait
+     * The number of epochs without improvement in the validation score to wait
      * before considering an early stop.
      *
      * @var int
@@ -354,7 +355,7 @@ class MLPRegressor implements Estimator, Online, Verbose, Persistable
         $bestScore = $min;
         $bestSnapshot = null;
         $prevLoss = INF;
-        $delta = 0;
+        $nu = 0;
 
         for ($epoch = 1; $epoch <= $this->epochs; $epoch++) {
             $batches = $training->randomize()->batch($this->batchSize);
@@ -382,9 +383,9 @@ class MLPRegressor implements Estimator, Online, Verbose, Persistable
                 $bestScore = $score;
                 $bestSnapshot = new Snapshot($this->network);
 
-                $delta = 0;
+                $nu = 0;
             } else {
-                $delta++;
+                $nu++;
             }
 
             if (is_nan($loss) or is_nan($score)) {
@@ -399,7 +400,7 @@ class MLPRegressor implements Estimator, Online, Verbose, Persistable
                 break 1;
             }
 
-            if ($delta >= $this->window) {
+            if ($nu >= $this->window) {
                 break 1;
             }